refactor(nav): move ls
